Removed AbstractTxReceipt moving useful functionality into AionTxReceipt
